refactor(observe): split wide event module
